Some type fixes.

+++ src/tools.cpp
@@ -461,7 +461,7 @@ void calculate_rmse(const mat_t& W_c, const mat_t& H_c, const char* srcdir, cons
 
     double nans_percentage = (double) nans_count / num_insts;
     printf("[INFO] NaNs percentage: %lf, NaNs Count: %d, Total Insts: %d\n", nans_percentage, nans_count, num_insts);
-    rmse = sqrtf(rmse / num_insts);
+    rmse = sqrt(rmse / num_insts);
     printf("[INFO] Test RMSE = %lf\n", rmse);
 }
